chat-with-pdf/requirements.txt
cassio
datasets
langchain
langchain_community
langchain_openai
openai
tiktoken
PyPDF2
python-dotenv
streamlit
